refactor: model 당사자가 dto의 필드 몰라도 되도록 수정
